Bugfix: allocate.{c,h} routines failed for larger arrays >2GB

+++ src/allocate.h
@@ -4,8 +4,8 @@
 
 #include <stdlib.h>
 
-void *get_spc(int num, size_t size);
-void *mget_spc(int num, size_t size);
+void *get_spc(size_t num, size_t size);
+void *mget_spc(size_t num, size_t size);
 void **get_img(int wd,int ht, size_t size);
 void ***get_3D(int N, int M, int A, size_t size);
 void free_img(void **pt);
